Модуляция создания постов
