+++ index.html
@@ -226,7 +226,6 @@
       </div>
     </section>
   </template>
-  <script src = "./js/functions.js"></script>
-  <script src = "./js/main.js"></script>
+  <script type="module" src = "./js/main.js"></script>
 </body>
 </html>
